pyproject.toml
[tool.poetry]
name = "waiter-menu"
version = "0.0.0"
description = ""
authors = ["parisa <parisafarivash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django = "^5.1"
channels = "^4.1.0"
websockets = "^12.0"
djangorestframework-simplejwt = "^5.3.1"
uvicorn = "^0.30.5"
pyjwt = "^2.9.0"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
